Mock db in auth adapter integration tests as well
