+++ tests/integration/auth_adapter/fixtures.py
@@ -20,18 +20,9 @@ from typing import Generator
 
 from fastapi.testclient import TestClient
 from pytest import fixture
-from testcontainers.mongodb import MongoDbContainer
 
 from auth_service.auth_adapter.api import main
-from auth_service.config import Config
-from auth_service.deps import (
-    UserDaoFactory,
-    get_config,
-    get_mongodb_config,
-    get_mongodb_dao_factory,
-    get_user_dao_factory,
-    get_user_dao_factory_config,
-)
+from auth_service.config import CONFIG
 
 
 @fixture(name="client")
@@ -40,36 +31,14 @@ def fixture_client() -> TestClient:
     return TestClient(main.app)
 
 
-@fixture(name="client_with_db")
-def fixture_client_with_db() -> Generator[
-    tuple[TestClient, UserDaoFactory], None, None
-]:
-    """Get test client for the auth adapter with a test database."""
-
-    with MongoDbContainer() as mongodb:
-        connection_url = mongodb.get_connection_url()
-        config = Config(db_url=connection_url, db_name="test-auth-adapter")
-
-        user_dao_factory = get_user_dao_factory(
-            config=get_user_dao_factory_config(config=config),
-            dao_factory=get_mongodb_dao_factory(
-                config=get_mongodb_config(config=config)
-            ),
-        )
-
-        main.app.dependency_overrides[get_config] = lambda: config
-        yield TestClient(main.app), user_dao_factory
-
-
 @fixture(name="with_basic_auth")
 def fixture_with_basic_auth() -> Generator[str, None, None]:
     """Run test with Basic authentication"""
     user, pwd = "testuser", "testpwd"
-    config = get_config()
-    config.basic_auth_user = user
-    config.basic_auth_pwd = pwd
+    CONFIG.basic_auth_user = user
+    CONFIG.basic_auth_pwd = pwd
     importlib.reload(main)
     yield f"{user}:{pwd}"
-    config.basic_auth_user = None
-    config.basic_auth_pwd = None
+    CONFIG.basic_auth_user = None
+    CONFIG.basic_auth_pwd = None
     importlib.reload(main)
