+++ tests/unit/auth_adapter/test_token_exchange.py
@@ -17,49 +17,13 @@
 """Unit tests for the core token exchange feature"""
 
 from datetime import datetime
-from typing import Any, Mapping, Optional
 
 from pytest import mark
 
 from auth_service.auth_adapter.core.auth import exchange_token
-from auth_service.user_management.models.dto import User, UserStatus
+from auth_service.user_management.models.dto import UserStatus
 
-from ...fixtures.utils import create_access_token, get_claims_from_token
-
-
-class DummyUserDao:
-    """UserDao that can retrieve one dummy user."""
-
-    def __init__(
-        self,
-        id_="john@ghga.org",
-        name="John Doe",
-        email="john@home.org",
-        ls_id="john@aai.org",
-    ):
-        """Initialize the dummy UserDao"""
-        self.user = User(
-            id=id_,
-            name=name,
-            email=email,
-            ls_id=ls_id,
-            status=UserStatus.ACTIVATED,
-            status_change=None,
-            registration_date=datetime(2020, 1, 1),
-        )
-
-    async def find_one(self, *, mapping: Mapping[str, Any]) -> Optional[User]:
-        """Find the dummy user."""
-        assert len(mapping) == 1
-        assert "ls_id" in mapping
-        if mapping["ls_id"] != self.user.ls_id:
-            return None
-        return self.user
-
-    async def update(self, user: User) -> None:
-        """Update the dummy user."""
-        assert user.id == self.user.id
-        self.user = user
+from ...fixtures.utils import DummyUserDao, create_access_token, get_claims_from_token
 
 
 @mark.asyncio
